+++ actionpack/lib/action_dispatch/railtie.rb
@@ -25,7 +25,6 @@ module ActionDispatch
     config.action_dispatch.use_cookies_with_metadata = false
     config.action_dispatch.perform_deep_munge = true
     config.action_dispatch.request_id_header = "X-Request-Id"
-    config.action_dispatch.return_only_request_media_type_on_content_type = true
     config.action_dispatch.log_rescued_responses = true
 
     config.action_dispatch.default_headers = {
@@ -51,7 +50,9 @@ module ActionDispatch
 
       ActiveSupport.on_load(:action_dispatch_request) do
         self.ignore_accept_header = app.config.action_dispatch.ignore_accept_header
-        self.return_only_media_type_on_content_type = app.config.action_dispatch.return_only_request_media_type_on_content_type
+        unless app.config.action_dispatch.respond_to?(:return_only_request_media_type_on_content_type)
+          self.return_only_media_type_on_content_type = app.config.action_dispatch.return_only_request_media_type_on_content_type
+        end
         ActionDispatch::Request::Utils.perform_deep_munge = app.config.action_dispatch.perform_deep_munge
       end