+++ railties/test/application/configuration_test.rb
@@ -3847,33 +3847,6 @@ module ApplicationTests
       assert_equal false, Rails.application.config.assets.unknown_asset_fallback
     end
 
-    test "ActionDispatch::Request.return_only_media_type_on_content_type is false by default" do
-      app "development"
-
-      assert_equal false, ActionDispatch::Request.return_only_media_type_on_content_type
-    end
-
-    test "ActionDispatch::Request.return_only_media_type_on_content_type is true in the 6.1 defaults" do
-      remove_from_config '.*config\.load_defaults.*\n'
-      add_to_config 'config.load_defaults "6.1"'
-
-      app "development"
-
-      assert_equal true, ActionDispatch::Request.return_only_media_type_on_content_type
-    end
-
-    test "ActionDispatch::Request.return_only_media_type_on_content_type can be configured in the new framework defaults" do
-      remove_from_config '.*config\.load_defaults.*\n'
-
-      app_file "config/initializers/new_framework_defaults_7_0.rb", <<-RUBY
-        Rails.application.config.action_dispatch.return_only_request_media_type_on_content_type = false
-      RUBY
-
-      app "development"
-
-      assert_equal false, ActionDispatch::Request.return_only_media_type_on_content_type
-    end
-
     test "action_dispatch.log_rescued_responses is true by default" do
       app "development"